Distances

The following distances measures are provided:

Distances induced in $l_p$ Spaces

VectorDataUtils.l2dist — Function

l2dist(a,b)

Compute the l2 distance between two real vectors a and b. The lengths of a and b must match.

VectorDataUtils.sql2dist — Function

sql2dist(a,b)

Compute the square of the l2 distance between two real vectors a and b. The lengths of a and b must match.

VectorDataUtils.l1dist — Function

l1dist(a,b)

Compute the l1 distance between two real vectors a and b. The lengths of a and b must match.

Mahalanobis Distance

VectorDataUtils.mahaldist — Function

mahaldist(x,y,A)

Compute the Mahalanobis distance between two real vectors x and y given a positive semidefinite matrix A. Dimensions of x, y and A must match. Mahalanobis distance is given by $sqrt((x-y)'A(x-y))$.