K-Means Clustering

K-means clustering of multidimensional vector data based on L2 distances is implemented. No specialised data structures are utilised and data is expected to be stored in matrices of Floats. The methods also do not check dimensions of inputs or bounds in for loops - it is expected that users will take care to avoid mistakes. Currently the emphasis is on speed (not convenience) with a view to implement GPU versions of these methods in the future.

Clustering Methods

VectorDataUtils.kmeans — Function

kmeans(data, number_of_centres; [update_threshold = 0.0])
kmeans(data, number_of_centres, initialise_method; [update_threshold = 0.0])

Estimate cluster centres from data using the k-means clustering algorithm. Cluster centres are iteratively updated until the largest shift in cluster centres, measured in terms of square of L2 distance, falls below update_threshold.

data must be a matrix of Float with each column representing a point

Default value of optional named parameter is update_threshold = 0.0 and corresponds to the setting that algorithm will converge when cluster centres do not shift

Use kmpp() as initialise_method to use k-mean++ as initialisation method [DEFAULT if initialise_method is not specified]

Use kmrand() as intialise_method to randomly select points as initial centres. This can be faster than k-means++ for large datasets

VectorDataUtils.kmeans! — Function

kmeans!(centres, data; [update_threshold = 0.0])

Updates (inplace) cluster centres centres using the k-means clustering algorithm based on data. Cluster centres are iteratively updated until the largest shift in cluster centres, measured in terms of square of L2 distance, falls below update_threshold.

centres must be a matrix of Float with each column representing a cluster centre

data must be a matrix of Float with each column representing a point

Default value of optional named parameter is update_threshold = 0.0 and corresponds to the setting that algorithm will converge when cluster centres do not shift

Support Methods

The following methods that support the kmeans algorithm implementation are also exposed to users

VectorDataUtils.kmeans_init — Function

kmeans_init([initialise_method], data, number_of_centres)

Pick a set of points from data to serve as the initial cluster centres for k-means clustering. Method for initialisation can random or k-means++.

data must be a matrix of Float with each column representing a point

Use kmpp() as initialise_method to use k-mean++ as initialisation method [DEFAULT if initialise_method is not specified]

Use kmrand() as intialise_method to randomly select points as initial centres. This can be much faster than k-means++ for large datasets

VectorDataUtils.closest_centre — Function

closest_centre(centres, data)

Identifies the centre closest to each data point based on L2 distance.

data must be a matrix of Float with each column representing a point

centres must be a matrix of Float with each column representing a centre

Output is a vector of Int with each element holding the column index of centres that corresponds to the closest centre

Note

closest_centre(centres, data) does not check the whether the dimensions of the input match. Users must ensure that size(data,1) = size(centres,1).

VectorDataUtils.closest_centre! — Function

closest_centre!(closest_centre, centres, data)

Identifies the centre closest to each data point based on L2 distance.

data must be a matrix of Float with each column representing a point

centres must be a matrix of Float with each column representing a centre

closest_centre must be a vector of Int with each element holding the column index of centres that corresponds to the closest centre

Note

closest_centre!(closest_centre, centres, data) does not check the whether the dimensions of the input match. Users must ensure that length(closest_centre) = size(data,2) and size(data,1) = size(centres,1).

VectorDataUtils.mindist2cntrs — Function

mindist2cntrs(centres, data)

Computes the square of the L2 distance to closest centre for each data point.

data must be a matrix of Float with each column representing a point

centres must be a matrix of Float with each column representing a centre

Note

mindist2cntrs(centres, data) does not check the whether the dimensions of the input match. Users must ensure that size(data,1) = size(centres,1).