Norms and Inner Products

The following elementary vector norms and inner products are provided:

$l_2$ Spaces

VectorDataUtils.l2inner — Function

l2inner(a,b)

Computes the inner product in the l2 vector space (dot product) between Real or Complex vectors a and b. The lenghts of a and b must match.

See also l2norm, l2dist, sql2norm, sql2dist.

Example

a = [1,2,3]
b = [2,0,4]
l2inner(a,b)

# output

14

l2inner(a)

Computes the inner product in the l2 vector space (dot product) of the Real or Complex vectors a with itself. i.e., the Squared l2 norm of a. This is identical to calling sql2norm(a).

VectorDataUtils.sql2norm — Function

sql2norm(a)

Compute the square of the l2 norm of a real valued vector a. This is identical to calling l2inner(a).

VectorDataUtils.l2norm — Function

l2norm(a)

Compute the l2 norm of a real valued vector a.

Other $l_p$ Spaces

VectorDataUtils.l1norm — Function

l1norm(a)

Compute the l1 norm of a real valued vector a.