Running Mean, Variance and Covariance

The following functions are provided to estimate the mean, covariance and variance along each dimension of a set of vectors. These are estimated as running statistics as per Welford, 1962.

Running Mean

VectorDataUtils.running_mean — Function

running_mean(x)

Compute the mean of a set of vectors x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

VectorDataUtils.running_mean! — Function

running_mean!(m, x)

Compute (in-place) the mean, m, of a set of vectors, x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

m must be a vector of Floats of appropriate length (equal to the number of rows in x).

VectorDataUtils.update_running_mean! — Function

update_running_mean!(m, x, n)

Update the mean, m, of some set of n-1 vectors to be the mean of the set of n vectors where x is the $n^th$ vector.

m and x must be vectors of equal length.

Typically used when the mean of a part of a dataset has been previously computed and needs to be updated with a new data point (vector).

Running Mean and Variance

VectorDataUtils.running_meanvar — Function

running_meanvar(x)

Compute the mean and variance along each dimension of a set of vectors, x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

VectorDataUtils.running_meanvar! — Function

running_meanvar!(m, v, x)

Compute (in-place) the mean, m, and variance along each dimension, v, of a set of vectors, x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

m must be a vector of Floats of appropriate length (equal to the number of rows in x).

v must be a vector of Floats of appropriate length (equal to the number of rows in x).

VectorDataUtils.update_running_meanvar! — Function

update_running_meanvar!(m, v, x, n)

Update the mean, m, and variance along each dimension, v, of some set of n-1 vectors to be the mean and variance of the set of n vectors where x is the $n^th$ vector (for n>=3).

m, v and x must be vectors of equal length.

Typically used when the mean and variance of a part of a dataset has been previously computed and needs to be updated with a new data point (vector).

Note

n must be 3 or greater. running_meanvar() or running_meanvar!() can be used to obtain an initial estimate of m and v from at least two vectors.

Running Mean and Covariance

VectorDataUtils.running_meancov — Function

running_meancov(x)

Compute the mean and covariance matrix a set of vectors, x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

VectorDataUtils.running_meancov! — Function

running_meancov!(m, C, x)

Compute (in-place) the mean, m, and covariance matrix, C, of a set of vectors, x.

x must be a matrix of Floats, with each column of x corresponding to a vector.

m must be a vector of Floats of appropriate length (equal to the number of rows in x).

C must be a square matrix of Floats of appropriate size (side equal to the number of rows in x).

VectorDataUtils.update_running_meancov! — Function

update_running_meancov!(m, C, x, n)

Update the mean, m, and covariance, C, of some set of n-1 vectors to be the mean and variance of the set of n vectors where x is the $n^th$ vector (for n>=3).

m and x must be vectors of equal length.

C must be a square matrix of side equal to the length of x.

Typically used when the mean and covariance of a part of a dataset has been previously computed and needs to be updated with a new data point (vector).

Note

n must be 3 or greater. running_meancov() or running_meanvcov!() can be used to obtain an initial estimate of m and v from at least two vectors.